% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCHN.R
\name{getCHN}
\alias{getCHN}
\title{Read in concordances from the Corpus Hedendaags Nederlands}
\usage{
getCHN(filename)
}
\arguments{
\item{filename}{The name of the file you read in. You can use file.choose() to choose a file interactively.}
}
\description{
This function allows for reading in concordances exported
from the free web interface for searching the Corpus Hedendaags Nederlands.
(You need a CLARIN account for using it, which many universities have.)
NOTE: There is currently no option to export the concordances themselves.
Just save the page with the query results in your browser and then read
in the HTML file.
}
