% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMAS.R
\name{getCOSMAS}
\alias{getCOSMAS}
\title{Read in concordances created with COSMAS II.}
\usage{
getCOSMAS(filename, merge = FALSE, years = FALSE,
  more_context = TRUE, ...)
}
\arguments{
\item{filename}{Path to the COSMAS II output file.}

\item{merge}{If matches consist of more than one word
(e.g. if you searched for something like "the X-er the Y-er" or
"the ADJ person" rather than just for e.g. "word"),
should they be merged in one column?}

\item{years}{Set to TRUE if you exported "Jahresansicht" or "Jahrzehntansicht".
Set to FALSE (the default) if you exported an unsorted concordance (which is
COSMAS II's default setting). The concordances package currently only supports
unsorted concordances, "Jahresansicht", and "Jahrzehntansicht".}

\item{more_context}{If TRUE (the default), the "Belege" section from the
concordance file is used to preserve the user-defined context window.
If FALSE, only the "KWIC" section from the concordance file is used,
which means that only ca. 5 words left and right are preserved,
ignoring the user-defined context window.}

\item{...}{arguments passed on from other methods}
}
\value{
A KWIC data frame.
}
\description{
Function for reading in concordances created with the search interface of the Institute for the German Language (IDS), COSMAS II, as data frames.
}
