% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCWB.R
\name{getCWB}
\alias{getCWB}
\title{Read in concordances created with the CWB Corpus Query Processor (CQP).}
\usage{
getCWB(filename, dt = TRUE)
}
\arguments{
\item{filename}{The name of the file you read in.}

\item{dt}{If TRUE (the default), the results will be returned as a data.table (see ?'data.table-package' for further information). If FALSE, they will be returned as a standard data frame.}
}
\description{
This function reads in export files created with the Corpus Query Processor (CQP) of the Corpus Workbench (CWB) as data.tables (from the data.table package) or normal R dataframes.
}
\examples{
~~ getCWB(myfile) # do not run
}
