% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWACKY.R
\name{getWACKY}
\alias{getWACKY}
\title{Read in concordances from the WaCkY Corpus}
\usage{
getWACKY(filename, tags = c("display", "omit", "column"), XML = TRUE)
}
\arguments{
\item{filename}{The name of the file you read in. You can use file.choose() to choose a file interactively. The file must be in XML (not .txt) format. It has to be formatted in UTF-8 (which should be the case by default).}

\item{tags}{string indicating how tokens and tags should be displayed, in case you exported a lemmatized or otherwise tagged concordance. "omit" will only show tokens and omit the lemmas. "display" will display the lemmas alongside the tokens in the same column, separated by /. "column" will add a "lemma" column to the concordance. "only" will only show lemmas in the keyword column. Default is "omit".}

\item{XML}{Is the concordance file in XML format? Default is TRUE. Currently only XML files are supported, support for .txt files will probably be implemented in the near future.}
}
\value{
A file containing a tab-separated concordance sheet.
}
\description{
This function allows for reading in concordances exported from the free web interface for searching the WaCkY corpora, http://nl.ijs.si/noske/wacs.cgi/first_form.
}
