% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export concordances as tab-separated files}
\usage{
export(x, filename, sep = "\\t", row.names = F,
  fileEncoding = "UTF-8", extension = "tsv", quote = F, ...)
}
\arguments{
\item{x}{The name of the object to be exported.}

\item{filename}{Path to the export file. By default, the file is named like the object and saved in the current working directory.}

\item{sep}{the field separator string. Values within each row of x are separated by this string}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written. Dafault is FALSE (i.e. no row names)}

\item{fileEncoding}{Intended encoding of the export file}

\item{extension}{File extension. Default is tsv. Ignored if a filename / file path is specified.}

\item{quote}{a logical value (TRUE or FALSE) or a numeric vector. If TRUE, any character or factor columns will be surrounded by double quotes. If a numeric vector, its elements are taken as the indices of columns to quote. In both cases, row and column names are quoted if they are written. If FALSE (the default), nothing is quoted.}

\item{...}{additional arguments to be passed on to write.table.}
}
\value{
A file containing a tab-separated concordance sheet.
}
\description{
This function - a convenient wrapper for write.table - exports R objects as tab-separated files (UTF-8 by default), which is usually the ideal format for corpus concordances.
}
