% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSE.R
\name{getNSE}
\alias{getNSE}
\title{Read in concordances created with NoSketchEngine (in particular, the NSE installation used by the COW corpora)}
\usage{
getNSE(filename, tags = TRUE)
}
\arguments{
\item{filename}{Path to the original concordance file.}

\item{tags}{Boolean: Does the concordance file contain tags (lemma, pos, ...)? Default is TRUE.}
}
\value{
A KWIC data frame or a KWIC csv sheet.
}
\description{
Function for reading in concordances created with NoSketchEngine.
}
